refactor: split off tracepoint
* move `Tracepoint` into its own module under core
